refactor: extract edit-interactive.ts, address reviewer feedback
- Extract runInteractiveEditLoop and all edit helpers (17 functions) to
  edit-interactive.ts so add.ts can import without cross-command dynamic import
- edit.ts: 1614 → 596 lines; edit-interactive.ts: 820 lines
- Rename generatedRouteToCurrentRoute → convertRouteToCurrentRoute
- Add comments explaining the retry loop purpose and 'manual' in test responses
- Replace dynamic import('./edit') in add.ts with static import from edit-interactive
